+++ Source/Libraries/Backend/Include/Backend/ShaderData/ShaderDataDescriptorInfo.h
@@ -5,5 +5,5 @@
 
 struct ShaderDataDescriptorInfo {
     /// Number of dwords within this descriptor data
-    size_t dwordCount{0};
+    uint32_t dwordCount{0};
 };
